#17077. Update PaymentOut to use Money instead of BigDecimal
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@24093 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ services/src/main/java/ish/oncourse/webservices/replication/v5/builders/PaymentOutStubBuilder.java
@@ -13,7 +13,7 @@ public class PaymentOutStubBuilder extends AbstractWillowStubBuilder<PaymentOut,
 		
 		PaymentOutStub stub = new PaymentOutStub();
 		
-		stub.setAmount(entity.getTotalAmount());
+		stub.setAmount(entity.getTotalAmount().toBigDecimal());
 		stub.setContactId(entity.getContact().getId());
 		stub.setCreated(entity.getCreated());
 		stub.setModified(entity.getModified());
